Move the MiniMapBtn and LayoutBtn to the new ControlCenter component which wraps and extends react-flow's Controls component
    ControlCenter's layout button toggle the layout direction (horizontal or vertical)

    Control Center renders a layout button

    minimap control button toggle map visibility

    Added benefit, this eliminates all the css associated with the previous control buttons

    render MiniMapBtn in out ControlCenter

    convert MiniMapBtn and LayoutBtn's from html native <button/> elements to react flow <ControlButton/> elements

    move ControlCenter component source tree to under the Tree component directory

    remove control buttons from Header
