--- src/components/Header/Controls/LayoutBtn/layoutbtn.module.css
@@ -1,28 +1,0 @@
-
-.layoutBtn {
-    margin: 0 0.25rem;
-    outline: 0;
-    cursor: pointer;
-    width: 2rem;
-    height: 2rem;
-    font-size: 1.2rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-weight: bold;
-    border-radius: 10px;
-    color: #334155;
-    background-color: #fafbfc;
-    border: 1px solid #1b1f2326;
-    box-shadow: rgba(27, 31, 35, 0.04) 0 1px 0 0, rgba(255, 255, 255, 0.25) 0px 1px 0px 0px inset;
-    transition: 0.2s cubic-bezier(0.3, 0, 0.5, 1);
-    transition-property: color, background-color, border-color;
-}
-
-.layoutBtn:hover {
-    color: #ffffff;
-    background-color: #334155;
-    border-color: #1b1f2326;
-    box-shadow: rgba(27, 31, 35, 0.1) 0 1px 0 0, rgba(255, 255, 255, 0.03) 0px 1px 0px 0px inset;
-    transition-duration: 0.1s;
-}
