+++ src/components/Header/header.module.css
@@ -29,8 +29,3 @@
     margin: 0;
 }
 
-.headerControls {
-    display: flex;
-    align-items: center;
-    justify-content: end;
-}
