+++ src/components/Tree/ControlCenter/index.ts
@@ -1,0 +1,1 @@
+export { ControlCenter } from 'components/Tree/ControlCenter/ControlCenter';
