+++ src/components/Tree/Tree.tsx
@@ -1,8 +1,9 @@
 import { BoolNode } from 'components/Nodes/BoolNode/BoolNode';
 import { DefaultNode } from 'components/Nodes/DefaultNode/DefaultNode';
-import { useDAG } from 'hooks';
-import React, { useMemo } from 'react';
-import ReactFlow, { Controls, Edge, MiniMap, NodeMouseHandler } from 'reactflow';
+import { ControlCenter } from 'components/Tree/ControlCenter';
+import { useDAG, useTreeDirection } from 'hooks';
+import React, { useMemo, useState } from 'react';
+import ReactFlow, { Edge, MiniMap, NodeMouseHandler } from 'reactflow';
 import { DagNode } from 'store/DagSlice/dagSlice';
 
 export interface TreeProps {
@@ -15,9 +16,11 @@ export interface TreeProps {
 /**
  * Tree - responsible for rendering the decision tree
  */
-export const Tree = ({ nodes, edges, onClick, mapVisible }: TreeProps) => {
+export const Tree = ({ nodes, edges, onClick }: TreeProps) => {
   const nodeTypes = useMemo(() => ({ BoolNode: BoolNode, default: DefaultNode }), []);
   const { onNodesChange, onEdgesChange } = useDAG();
+  const [mapVisible, setMapVisible] = useState(true);
+  const [direction, setDirection] = useTreeDirection();
 
   return (
     <>
@@ -34,7 +37,12 @@ export const Tree = ({ nodes, edges, onClick, mapVisible }: TreeProps) => {
           proOptions={{ hideAttribution: true }}
         >
           {mapVisible && <MiniMap nodeStrokeWidth={3} data-testid="tree-mini-map" />}
-          <Controls showInteractive={false} />
+          <ControlCenter
+            mapVisible={mapVisible}
+            setMapVisible={setMapVisible}
+            direction={direction}
+            setDirection={setDirection}
+          />
         </ReactFlow>
       </div>
     </>
